+++ packages/bpk-component-calendar/index.js
@@ -4,6 +4,7 @@ import BpkCalendarNav from './src/BpkCalendarNav';
 import BpkCalendarDate from './src/BpkCalendarDate';
 import BpkCalendarView from './src/BpkCalendarView';
 import CustomPropTypes from './src/custom-proptypes';
+import * as DateUtils from './src/date-utils';
 
 export default BpkCalendar;
 export {
@@ -12,4 +13,5 @@ export {
   BpkCalendarDate,
   BpkCalendarView,
   CustomPropTypes,
+  DateUtils,
 };